Clean up API header files
