+++ libattachsql-1.0/attachsql.h
@@ -29,10 +29,11 @@ typedef uint16_t in_port_t;
 # include <arpa/inet.h>
 #endif
 
-/* These three need to go first, in this order */
+/* These four need to go first, in this order */
 #include <libattachsql-1.0/visibility.h>
-#include <libattachsql-1.0/error.h>
 #include <libattachsql-1.0/constants.h>
+#include <libattachsql-1.0/structs.h>
+#include <libattachsql-1.0/error.h>
 
 #include <libattachsql-1.0/connect.h>
 #include <libattachsql-1.0/group.h>
