+++ libattachsql-1.0/query.h
@@ -21,85 +21,6 @@
 extern "C" {
 #endif
 
-enum attachsql_column_flags_t
-{
-  ATTACHSQL_COLUMN_FLAGS_NONE=              0,
-  ATTACHSQL_COLUMN_FLAGS_NOT_NULL=          (1 << 0),
-  ATTACHSQL_COLUMN_FLAGS_PRIMARY_KEY=       (1 << 1),
-  ATTACHSQL_COLUMN_FLAGS_UNIQUE_KEY=        (1 << 2),
-  ATTACHSQL_COLUMN_FLAGS_MULTIPLE_KEY=      (1 << 3),
-  ATTACHSQL_COLUMN_FLAGS_BLOB=              (1 << 4), // Deprecated in MySQL
-  ATTACHSQL_COLUMN_FLAGS_UNSIGNED=          (1 << 5),
-  ATTACHSQL_COLUMN_FLAGS_ZEROFILL=          (1 << 6),
-  ATTACHSQL_COLUMN_FLAGS_BINARY=            (1 << 7),
-  ATTACHSQL_COLUMN_FLAGS_ENUM=              (1 << 8),
-  ATTACHSQL_COLUMN_FLAGS_AUTO_INCREMENT=    (1 << 9),
-  ATTACHSQL_COLUMN_FLAGS_TIMESTAMP=         (1 << 10), // Deprecated in MySQL
-  ATTACHSQL_COLUMN_FLAGS_SET=               (1 << 11),
-  ATTACHSQL_COLUMN_FLAGS_NO_DEFAULT_VALUE=  (1 << 12),
-  ATTACHSQL_COLUMN_FLAGS_ON_UPDATE_NOW=     (1 << 13), // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_PART_KEY=          (1 << 14), // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_NUM=               (1 << 15),
-  ATTACHSQL_COLUMN_FLAGS_GROUP=             (1 << 15), // NUM and GROUP are both this // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_UNIQUE=            (1 << 16), // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_BINCMP=            (1 << 17), // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_GET_FIELD_COLUMNS= (1 << 18), // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_IN_PART_FUNC=      (1 << 19), // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_IN_ADD_INDEX=      (1 << 20), // Undocumented in MySQL
-  ATTACHSQL_COLUMN_FLAGS_RENAMED=           (1 << 21)  // Undocumented in MySQL
-};
-
-typedef enum attachsql_column_flags_t attachsql_column_flags_t;
-
-enum attachsql_query_parameter_type_t
-{
-  ATTACHSQL_ESCAPE_TYPE_NONE,
-  ATTACHSQL_ESCAPE_TYPE_CHAR,
-  ATTACHSQL_ESCAPE_TYPE_CHAR_LIKE,
-  ATTACHSQL_ESCAPE_TYPE_INT,
-  ATTACHSQL_ESCAPE_TYPE_BIGINT,
-  ATTACHSQL_ESCAPE_TYPE_FLOAT,
-  ATTACHSQL_ESCAPE_TYPE_DOUBLE
-};
-
-typedef enum attachsql_query_parameter_type_t attachsql_query_parameter_type_t;
-
-struct attachsql_query_parameter_st
-{
-  attachsql_query_parameter_type_t type;
-  void *data;
-  size_t length;
-  bool is_unsigned;
-};
-
-typedef struct attachsql_query_parameter_st attachsql_query_parameter_st;
-
-struct attachsql_query_column_st
-{
-  char *schema;
-  char *table;
-  char *origin_table;
-  char *column;
-  char *origin_column;
-  uint16_t charset;
-  uint32_t length;
-  attachsql_column_type_t type;
-  attachsql_column_flags_t flags;
-  uint8_t decimals;
-  char *default_value;
-  size_t default_size;
-};
-
-typedef struct attachsql_query_column_st attachsql_query_column_st;
-
-struct attachsql_query_row_st
-{
-  char *data;
-  size_t length;
-};
-
-typedef struct attachsql_query_row_st attachsql_query_row_st;
-
 ASQL_API
 bool attachsql_query(attachsql_connect_t *con, size_t length, const char *statement, uint16_t parameter_count, attachsql_query_parameter_st *parameters, attachsql_error_t **error);
